agentic_chatbot_RAG/pyproject.toml
[tool.poetry]
name = "agentic-chatbot-rag"
version = "0.1.0"
description = ""
authors = ["Sanjeev <nsanjeevram@outlook.com>"]
readme = "README.md"
packages = [{include = "agentic_chatbot_rag", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
llama-index-core = "0.10.30"
llama-index-program-evaporate = "^0.1.2"
llama-index-llms-gemini = "0.1.7"
llama-index-embeddings-gemini = "0.1.6"
streamlit = "^1.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
